stage2/06-services-config/files/requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
psutil==7.0.0
pyrtlsdr==0.3.0
